requirements.txt
# OpenTelemetry core packages
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Azure Monitor exporters for Application Insights
azure-monitor-opentelemetry-exporter>=1.0.0b21

# Optional: Automatic instrumentation for common libraries
opentelemetry-instrumentation>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-flask>=0.41b0

# Python standard library (no additional deps needed)
